Enterprise asset tracking by hostname
The purpose of this little tool is to track enterprise assets by hostname
for the major purpose of detecting rouge devices. The input data can be
any logs that relate to an asset by hostname, for example:

  - {your vendor} EDR logs
  - Azure AD device logs
  - {your vendor} Anti-Virus logs
  - Windows logs

You can also configure requirements as to what an asset has to have
as a recently logged security tool (such as your EDR and AV clients).